Website Update
Created the makeTree function that creates the node tree for the source article on the page. It sets up and places the node tree within the HTML document and displays the node counts from the document. Also created the makeBranches function that updates the contents of the node tree and the count of the different global variables, and appends node branches to the node tree diagram.